WIP #2429 Updated conventions and annotations for OceanData service
