+++ test/unit/server/rpc/procedures/ocean-data.spec.js
@@ -1,0 +1,13 @@
+describe('ocean-data', function() {
+    const utils = require('../../../../assets/utils');
+    var OceanData = utils.reqSrc('rpc/procedures/ocean-data/ocean-data'),
+        RPCMock = require('../../../../assets/mock-rpc'),
+        oceanData = new RPCMock(OceanData);
+
+    utils.verifyRPCInterfaces(oceanData, [
+        ['getOxygenRatio', []],
+        ['getDeepOceanTemp', []],
+        ['getSurfaceTemp', []],
+        ['getSeaLevel', []],
+    ]);
+});
