--- test/unit/server/rpc/procedures/sea-surface.spec.js
@@ -1,14 +1,0 @@
-describe('sea-surface', function() {
-    const utils = require('../../../../assets/utils');
-    var OceanData = utils.reqSrc('rpc/procedures/sea-surface/sea-surface'),
-        RPCMock = require('../../../../assets/mock-rpc'),
-        seaSurface = new RPCMock(OceanData);
-
-    utils.verifyRPCInterfaces(seaSurface, [
-        ['getFullTable', []],
-        ['getOxygenRatio', []],
-        ['getDeepOceanTemp', []],
-        ['getSurfaceTemp', []],
-        ['getSeaLevel', []],
-    ]);
-});
